requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2.7
django-cleanup==5.2.0
django-faker==0.2
django-widget-tweaks==1.4.8
fake-factory==9999.9.9
fpdf==1.7.2
idna==3.2
numpy==1.21.2
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.60.0
urllib3==1.26.7
